theater-master/requirements.txt
Django==1.11.7
django-filter==1.0.4
django-ordered-model==1.4.3
django-rest-swagger==2.1.2
invoke==0.21.0
Pillow==4.2.1
pytz==2017.3
uWSGI==2.0.15
